+++ packages/lumx-core/src/js/types.ts
@@ -1,0 +1,6 @@
+import { GLOBAL_THEMES } from '@lumx/core/js/constants';
+
+/**
+ * Global theme (lumapps/material).
+ */
+export type GlobalTheme = keyof typeof GLOBAL_THEMES;