+++ packages/lumx-core/style-dictionary/config/utils/_prettier-ts.js
@@ -1,0 +1,11 @@
+const StyleDictionary = require('style-dictionary');
+const execSync = require('child_process').execSync;
+
+module.exports = function ({ buildPath }) {
+    const name = 'prettier/ts';
+    StyleDictionary.registerAction({
+        name,
+        do: () => execSync(`yarn prettier-tslint fix '${buildPath}'`),
+    });
+    return name;
+};